Take a copy of one of the AWS Glue scripts
This will take load the data from an S3 file, transform it and spit it out into both Parquet and Postgres.

The transformation at the moment doesn't actually change the structure of the data so the mapping should be a like-for-like. This would change in the future to add and remove any fields we like.