Query offices by proximity
This depends on the CSV data being changed before it's pulled. The ID column on each file needs to become `geolocation_foreign_key` (or another name we chose) so that GIS indexing works.

Next we need to think about returning 10 offices being ok or does it need to be from 10 miles of the current location?