Clean office data
`liberal_parsing` is an important option to have set as Ruby was doing odd things to the new CSV which Postgres COPY then couldn't read from.

We know that rows in the office CSV should contain all these fields and any that don't include ids and long/lats should have been filtered out upstream. For now we ignore any rows that are missing these values.

Some cells within the columns used for free text ended up with escaped double quotes:

"Hello my name is \"Frank\"." When creating the cleaned CSV Ruby was upset and didn't interpret the quote correctly. I swap them to single quotes and all is well.

If we don't have to clean the data for missing rows in future we might be able to skip this entirely and give the CSV straight to Postgres COPY.