Update office schema
Now we know that long and lats are expected we can make them required again.

The field 'about_our_advice_service__c' was spelt wrong so we correct that

With an eye on the future we will probably want to look up LA via the parentid field, so we make that mandatory too.